+++ client/src/components/UserStatsGrid.module.scss
@@ -1,0 +1,49 @@
+.statsGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  gap: 0.5rem;
+  margin-top: 1rem;
+  width: 100%;
+  grid-template-areas:
+    "all all all"
+    "easy medium hard";
+  color: #9C9995;
+
+  & > div {
+    background-color: #2C2C2C;
+    height: 3.5rem;
+    border-radius: 0.5rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 0.125rem;
+
+    h3 {
+      margin: 0;
+      padding: 0;
+      font-size: 0.75rem;
+    }
+
+    p {
+      color: #fff;
+      font-weight: 600;
+    }
+
+    &:nth-child(1) {
+      grid-area: all;
+    }
+
+    &:nth-child(2) {
+      grid-area: easy;
+    }
+
+    &:nth-child(3) {
+      grid-area: medium;
+    }
+
+    &:nth-child(4) {
+      grid-area: hard;
+    }
+  }
+}
