+++ client/src/pages/DashboardPage.module.scss
@@ -31,55 +31,6 @@
     justify-items: center;
     text-align: center;
     grid-area: stats;
-
-    .statsGrid {
-      display: grid;
-      grid-template-columns: 1fr 1fr 1fr;
-      gap: 0.5rem;
-      margin-top: 1rem;
-      width: 100%;
-      grid-template-areas:
-        "all all all"
-        "easy medium hard";
-
-      & > div {
-        background-color: #2C2C2C;
-        height: 3.5rem;
-        border-radius: 0.5rem;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        gap: 0.125rem;
-
-        h3 {
-          margin: 0;
-          padding: 0;
-          font-size: 0.75rem;
-        }
-
-        p {
-          color: #fff;
-          font-weight: 600;
-        }
-
-        &:nth-child(1) {
-          grid-area: all;
-        }
-
-        &:nth-child(2) {
-          grid-area: easy;
-        }
-
-        &:nth-child(3) {
-          grid-area: medium;
-        }
-
-        &:nth-child(4) {
-          grid-area: hard;
-        }
-      }
-    }
   }
 
   .feed, .relationships {
@@ -94,14 +45,6 @@
 
   .relationships {
     grid-area: relationships;
-  }
-
-  @media screen and (min-width: 768px) {
-    grid-template-columns: 1fr 1fr;
-    grid-template-rows: 1fr 1fr 1fr;
-    grid-template-areas: 
-      "stats feed"
-      "relationships feed";
   }
 
   @media screen and (min-width: 1200px) {
